ashrae/requirements.txt
alabaster==0.7.12
astroid==2.3.3
awscli==1.16.303
Babel==2.7.0
botocore==1.13.39
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
cycler==0.10.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
idna==2.8
imagesize==1.1.0
isort==4.3.21
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lightgbm==2.3.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
torch==1.3.1
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
